+++ Sources/SwiftletModel/Relation/Relation+Codable.swift
@@ -1,0 +1,217 @@
+//
+//  File.swift
+//  
+//
+//  Created by Sergey Kazakov on 04/08/2024.
+//
+
+import Foundation
+
+// MARK: - Codable
+
+extension Relation: Codable where Entity: Codable {
+    public func encode(to encoder: any Encoder) throws {
+        switch encoder.relationEncodingStrategy {
+        case .plain:
+            try encodePlainContainer(to: encoder)
+        case .keyedContainer:
+            try encodeKeyedContainer(to: encoder)
+        case .explicitKeyedContainer:
+            try encodeExplicitKeyedContainer(to: encoder)
+        }
+    }
+
+    public init(from decoder: any Decoder) throws {
+        if decoder.relationDecodingStrategy.contains(.explicitKeyedContainer),
+           let relation = try? Self.decodeExplicitKeyedContainer(from: decoder) {
+
+            self = relation
+            return
+        }
+
+        if decoder.relationDecodingStrategy.contains(.keyedContainer),
+           let relation = try? Self.decodeKeyedContainer(from: decoder) {
+
+            self = relation
+            return
+        }
+
+        self = try Self.decodePlainContainer(from: decoder)
+    }
+}
+
+// MARK: - Codable Explicitly
+
+extension Relation where Entity: Codable {
+    enum RelationCodingKeys: String, CodingKey {
+        case id = "id"
+        case entity = "object"
+        case ids = "ids"
+        case entities = "objects"
+        case none
+    }
+
+    func encodeKeyedContainer(to encoder: Encoder) throws {
+        switch state {
+        case .id(let value):
+            var container = encoder.container(keyedBy: RelationCodingKeys.self)
+            try container.encode(value, forKey: .id)
+        case .entity(let value, _):
+            var container = encoder.container(keyedBy: RelationCodingKeys.self)
+            try container.encode(value, forKey: .entity)
+        case .ids(let value, _):
+            var container = encoder.container(keyedBy: RelationCodingKeys.self)
+            try container.encode(value, forKey: .ids)
+        case .entities(let value, _, _):
+            var container = encoder.container(keyedBy: RelationCodingKeys.self)
+            try container.encode(value, forKey: .entities)
+        case .none:
+            var container = encoder.singleValueContainer()
+            try container.encodeNil()
+        }
+    }
+
+    static func decodeKeyedContainer(from decoder: any Decoder) throws -> Relation {
+        guard let container = try? decoder.container(keyedBy: RelationCodingKeys.self),
+              let key = container.allKeys.first
+        else {
+            return .none
+        }
+
+        switch key {
+        case .id:
+            let value = try? container.decode(Entity.ID?.self, forKey: .id)
+            return Relation(state: .id(id: value))
+        case .entity:
+            let value = try? container.decode(Entity?.self, forKey: .entity)
+            return Relation(state: .entity(entity: value, fragment: false))
+        case .ids:
+            let value = try container.decode([Entity.ID].self, forKey: .ids)
+            return Relation(state: .ids(ids: value, chunk: false))
+        case .entities:
+            let value = try container.decode([Entity].self, forKey: .entities)
+            return Relation(state: .entities(entities: value, chunk: false, fragment: false))
+        case .none:
+            return .none
+        }
+    }
+}
+
+// MARK: - Codable Exactly
+
+extension Relation where Entity: Codable {
+    enum RelationExplicitCodingKeys: String, CodingKey {
+        case id = "id"
+        case entity = "object"
+        case ids = "ids"
+        case entities = "objects"
+        case idsChunk = "chunk_ids"
+        case chunk = "chunk"
+        case none
+    }
+
+    func encodeExplicitKeyedContainer(to encoder: Encoder) throws {
+        switch state {
+        case .id(let value):
+            var container = encoder.container(keyedBy: RelationExplicitCodingKeys.self)
+            try container.encode(value, forKey: .id)
+        case .entity(let value, _):
+            var container = encoder.container(keyedBy: RelationExplicitCodingKeys.self)
+            try container.encode(value, forKey: .entity)
+        case .ids(let value, let chunk):
+            var container = encoder.container(keyedBy: RelationExplicitCodingKeys.self)
+            chunk ?
+            try container.encode(value, forKey: .idsChunk)
+            : try container.encode(value, forKey: .ids)
+           
+        case .entities(let value, let chunk, _):
+            var container = encoder.container(keyedBy: RelationExplicitCodingKeys.self)
+            chunk ?
+            try container.encode(value, forKey: .chunk)
+            : try container.encode(value, forKey: .entities)
+        case .none:
+            var container = encoder.singleValueContainer()
+            try container.encodeNil()
+        }
+    }
+
+    static func decodeExplicitKeyedContainer(from decoder: any Decoder) throws -> Relation {
+        guard let container = try? decoder.container(keyedBy: RelationExplicitCodingKeys.self),
+              let key = container.allKeys.first
+        else {
+            return .none
+        }
+
+        switch key {
+        case .id:
+            let value = try? container.decode(Entity.ID?.self, forKey: .id)
+            return Relation(state: .id(id: value))
+        case .entity:
+            let value = try? container.decode(Entity?.self, forKey: .entity)
+            return Relation(state: .entity(entity: value, fragment: false))
+        case .ids:
+            let value = try container.decode([Entity.ID].self, forKey: .ids)
+            return Relation(state: .ids(ids: value, chunk: false))
+        case .entities:
+            let value = try container.decode([Entity].self, forKey: .entities)
+            return Relation(state: .entities(entities: value, chunk: false, fragment: false))
+        case .idsChunk:
+            let value = try container.decode([Entity.ID].self, forKey: .idsChunk)
+            return Relation(state: .ids(ids: value, chunk: true))
+        case .chunk:
+            let value = try container.decode([Entity].self, forKey: .chunk)
+            return Relation(state: .entities(entities: value, chunk: true, fragment: false))
+        case .none:
+            return .none
+        }
+    }
+}
+
+// MARK: - Codable Flattaned
+
+extension Relation where Entity: Codable {
+    // swiftlint:disable:next type_name
+    struct ID<T: EntityModelProtocol>: Codable {
+        let id: T.ID
+    }
+
+    func encodePlainContainer(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        switch state {
+        case .id(let value):
+            try container.encode(value.map { ID<Entity>(id: $0) })
+        case .entity(let value, _):
+            try container.encode(value)
+        case .ids(let value, _):
+            try container.encode(value.map { ID<Entity>(id: $0) })
+        case .entities(let value, _, _):
+            try container.encode(value)
+        case .none:
+            try container.encodeNil()
+        }
+    }
+
+    static func decodePlainContainer(from decoder: any Decoder) throws -> Relation {
+        guard let container = try? decoder.singleValueContainer() else {
+            return .none
+        }
+
+        if let value = try? container.decode(Entity?.self) {
+            return Relation(state: .entity(entity: value, fragment: false))
+        }
+
+        if let value = try? container.decode(ID<Entity>?.self) {
+            return Relation(state: .id(id: value.id))
+        }
+
+        if let value = try? container.decode([Entity].self) {
+            return Relation(state: .entities(entities: value, chunk: false, fragment: false))
+        }
+
+        if let value = try? container.decode([ID<Entity>].self) {
+            return Relation(state: .ids(ids: value.map { $0.id }, chunk: false))
+        }
+
+        return .none
+    }
+}
